+++ core/employee_roles/serializers.py
@@ -1,4 +1,4 @@
-from step.models import EmployeeRole
+from core.models import EmployeeRole
 from rest_framework import serializers
 
 class EmployeeRoleSerializer(serializers.HyperlinkedModelSerializer):
