+++ core/participants/serializers.py
@@ -1,4 +1,4 @@
-from step.models import Participant
+from core.models import Participant
 from rest_framework import serializers
 
 class ParticipantSerializer(serializers.HyperlinkedModelSerializer):
